Split Instance.hs on Instance.hs and Process.hs
